requirements.txt
python-docx
python-dotenv
pandas